Allow 'Extract fields' transformer to extract into a single field
I've been playing around with #41994 and found that it requires fields
to contain array values, which can't be sent from a backend plugin. This
PR adds the ability for the `ExtractFieldsTransformer` to extract values
from a field into a complex object in a single field, rather than into
multiple separate fields. A field containing a JSONified array can then
be transformed into a field containing an actual array, which can in
turn be used for the table charts in #41994.

I originally added a whole separate `ParseJSONTransformer` which did
something similar, before finding that the `ExtractFieldsTransformer`
already had some JSON parsing responsibility. If this adds too much
complexity to `ExtractFieldsTransformer` I can submit a PR for the
separate transformer instead, though.

I'll add tests later if we decide to proceed with this PR!

(Alternatively, if there's already a way of creating an array field that
I'm missing, I'm happy to abandon this...)